functions/fetch/src/requirements.txt
beautifulsoup4==4.8.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-pubsub==0.45.0
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.12.3
grpcio==1.22.0
html2text==2019.8.11
idna==2.8
jieba3k==0.35.1
lxml==4.4.1
markdown2==2.3.8
newspaper3k==0.2.8
nltk==3.4.4
Pillow==6.1.0
protobuf==3.9.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-file==1.4.3
rsa==4.0
ruamel.yaml==0.16.1
ruamel.yaml.clib==0.1.2
six==1.12.0
soupsieve==1.9.2
tinysegmenter==0.3
tldextract==2.2.1
urllib3==1.25.3
